requirements.txt
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
idna==2.10
numpy==1.19.5
pandas==1.2.1
pycparser==2.20
pycryptodome==3.10.1
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.23
urllib3==1.26.3
xlrd==2.0.1
redis==3.5.3
beautifulsoup4==4.9.3
plotly==5.1.0
scikit-learn==0.24.1